works on 64bit (i think) and lapack is pulled in by "extern"
git-svn-id: file:///home/mb312/repositories/public-repo/bic/trunk@1309 c5881a9f-4ef2-0310-b211-a1cdbd18b07b

+++ misc_tools/auto_shim/image.h
@@ -5,7 +5,7 @@ typedef struct{
   short int  status;       // status word for block header
   short int  index;     // spare short word
   short int  mode;     // spare short word
-  long int   ctcount;     // spare long word
+  int   ctcount;     // spare long word
   float      lpval1;       // 2D-f2 left phase
   float      rpval1;       // 2D-f2 right phase
   float      lvl;     // spare float word
